+++ flatlaf-intellij-themes/src/main/java/com/formdev/flatlaf/intellijthemes/materialthemeuilite/FlatMTSolarizedLightIJTheme.java
@@ -26,25 +26,25 @@ import com.formdev.flatlaf.IntelliJTheme;
 /**
  * @author Karl Tauber
  */
-public class FlatMonokaiProIJTheme
+public class FlatMTSolarizedLightIJTheme
 	extends IntelliJTheme.ThemeLaf
 {
-	public static final String NAME = "Monokai Pro (Material)";
+	public static final String NAME = "Solarized Light (Material)";
 
 	public static boolean setup() {
 		try {
-			return setup( new FlatMonokaiProIJTheme() );
+			return setup( new FlatMTSolarizedLightIJTheme() );
 		} catch( RuntimeException ex ) {
 			return false;
 		}
 	}
 
 	public static void installLafInfo() {
-		installLafInfo( NAME, FlatMonokaiProIJTheme.class );
+		installLafInfo( NAME, FlatMTSolarizedLightIJTheme.class );
 	}
 
-	public FlatMonokaiProIJTheme() {
-		super( Utils.loadTheme( "Monokai Pro.theme.json" ) );
+	public FlatMTSolarizedLightIJTheme() {
+		super( Utils.loadTheme( "Solarized Light.theme.json" ) );
 	}
 
 	@Override